First draft of verhuis service
