+++ .env
@@ -9,9 +9,9 @@
 ##################################################
 
 # The shortcode for this component, should be a small set of letters reprecentint the application
-APP_NAME=ts
+APP_NAME=vs
 # The Full title of the application
-APP_TITLE=Trouw Service
+APP_TITLE=Verhuis Service
 # The current version of the application
 APP_VERSION=V.0.1
 # Do you want to display the symfony debug toolbar?
@@ -30,8 +30,8 @@ APP_ID=
 # The type of component, can either be application or component
 APP_TYPE=component
 
-APP_LOGO=pc.zaakonline.nl
-APP_HOME=pc.zaakonline.nl
+APP_LOGO=vs.zuid-drecht.nl
+APP_HOME=vs.zuid-drecht.nl
 
 # The urls on wich this api is available
 TRUSTED_PROXIES=10.0.0.0/8,172.16.0.0/12,192.168.0.0/16
@@ -73,31 +73,32 @@ ORGANIZATION_UNIT_NAME=Common-Ground
 ##################################################
 
 # The domain for this component (do not include a subdomain, this will hamper certification generation)
-APP_DOMAIN=conduction.nl
+APP_DOMAIN=zuid-drecht.nl
 
 # The subdomain for this component (should be false if you want to use www)
-APP_SUBDOMAIN=ts
+APP_SUBDOMAIN=vs
 
 # The demo enviroment for this component @depracticed
-APP_DEMO=ts.zaakonline.nl
+APP_DEMO=vs.zuid-drecht.nl
 
 # The Repository for this component
-APP_REPRO=https://github.com/ConductionNL/trouw-service
+APP_REPRO=https://github.com/ConductionNL/verhuis-service
 
 # The Helm Charts for this components
-APP_HELM=https://github.com/ConductionNL/trouw-service/helm
+APP_HELM=https://github.com/ConductionNL/verhuis-service/helm
 
 # The repository for the primary (php) container of this project
-APP_CONTAINER=https://hub.docker.com/repository/docker/conduction/ts-php
+APP_CONTAINER=https://hub.docker.com/repository/docker/conduction/vs-php
 
 ##################################################
 # Docker settings
 ##################################################
 
 CONTAINER_REGISTRY_BASE=docker.io/conduction
-CONTAINER_PROJECT_NAME=ts
+CONTAINER_PROJECT_NAME=vs
+
 # The repository for the primary (php) container of this project
-CONTAINER_REPRO=https://hub.docker.com/repository/docker/conduction/ts-php
+CONTAINER_REPRO=https://hub.docker.com/repository/docker/conduction/vs-php
 
 ##################################################
 # ZGW settings
