+++ api/helm/Chart.yaml
@@ -1,7 +1,7 @@
 apiVersion: v1
 appVersion: V.0.1
 description: '&#039;Naast deze JSON rest API is er ook een [graphql](/graphql) interface beschikbaar.&#039;'
-name: ts
+name: vs
 version: 0.1.0
-home: pc.zaakonline.nl
-icon: pc.zaakonline.nl
+home: vs.zuid-drecht.nl
+icon: vs.zuid-drecht.nl
