+++ api/public/schema/openapi.yaml
@@ -1,16 +1,16 @@
 openapi: 3.0.2
 info:
-  title: 'Trouw Service'
+  title: 'Verhuis Service'
   version: V.0.1
   description: |
     API Details
-    - Component: Trouw Service
-    - Reference: ts
+    - Component: Verhuis Service
+    - Reference: vs
     - Enviroment: dev
     - Version: V.0.1
-    - Repository: [online](https://github.com/ConductionNL/trouw-service) / [zip](https://github.com/ConductionNL/trouw-service/archive/master.zip)
-    - Docker Image: [online](https://hub.docker.com/repository/docker/conduction/ts-php) 
-    - Datamodel: [postman](/schema/ts.postman_collection.json) / [oas](/schema/openapi.yaml) / [pdf](/schema/datamodel.pdf) / [modelio (zip)](/schema/Datamodel_Modelio.zip) 
+    - Repository: [online](https://github.com/ConductionNL/verhuis-service) / [zip](https://github.com/ConductionNL/verhuis-service/archive/master.zip)
+    - Docker Image: [online](https://hub.docker.com/repository/docker/conduction/vs-php) 
+    - Datamodel: [postman](/schema/vs.postman_collection.json) / [oas](/schema/openapi.yaml) / [pdf](/schema/datamodel.pdf) / [modelio (zip)](/schema/Datamodel_Modelio.zip) 
     
     'Naast deze JSON rest API is er ook een [graphql](/graphql) interface beschikbaar.'         
     
@@ -3441,12 +3441,12 @@ definitions:
         description: 'The moment this request was created'
         type: string
         format: date-time
-        example: '2020-06-15 09:40:32'
+        example: '2020-06-26 05:26:03'
       dateModified:
         description: 'The moment this request last Modified'
         type: string
         format: date-time
-        example: '2020-06-15 09:40:32'
+        example: '2020-06-26 05:26:03'
     required: []
   AuditTrail-write:
     properties:
@@ -3494,12 +3494,12 @@ definitions:
         description: 'The moment this request was created'
         type: string
         format: date-time
-        example: '2020-06-15 09:40:32'
+        example: '2020-06-26 05:26:03'
       dateModified:
         description: 'The moment this request last Modified'
         type: string
         format: date-time
-        example: '2020-06-15 09:40:32'
+        example: '2020-06-26 05:26:03'
     required: []
   ChangeLog-write:
     properties:
@@ -3546,8 +3546,8 @@ securityDefinitions:
 host: irc.zaakonline.nl
 servers:
   -
-    name: conduction.nl
-    url: ts.dev.conduction.nl
+    name: zuid-drecht.nl
+    url: vs.dev.zuid-drecht.nl
 externalDocs:
   -
     description: 'OAS Documentatie'
