+++ api/public/schema/publiccode.yaml
@@ -1,13 +1,13 @@
 publiccodeYmlVersion: "0.2"
 
-name: ts
+name: vs
 applicationSuite: commonground
-url: "https://github.com/ConductionNL/trouw-service"
-landingURL: "ts.zaakonline.nl" 
+url: "https://github.com/ConductionNL/verhuis-service"
+landingURL: "vs.zuid-drecht.nl" 
 isBasedOn: "https://github.com/ConductionNL/Proto-component-commonground.git" 
 softwareVersion: "V.0.1"
-releaseDate: "2020-15-06"
-logo: pc.zaakonline.nl 
+releaseDate: "2020-26-06"
+logo: vs.zuid-drecht.nl 
 monochromeLogo: img/logo-mono.svg
 
 inputTypes:
@@ -37,7 +37,7 @@ usedBy:
   - Gemeente Hoorn
   
 
-roadmap: "https://github.com/ConductionNL/trouw-service/blob/master/ROADMAP.md"
+roadmap: "https://github.com/ConductionNL/verhuis-service/blob/master/ROADMAP.md"
 
 developmentStatus: development
 
@@ -66,8 +66,8 @@ description:
 #          on multiple rows. You should note what the software
 #          is and why one should need it.
 
-    documentation: "https://github.com/ConductionNL/trouw-service/blob/master/README.md"
-    apiDocumentation: "ts.zaakonline.nl"
+    documentation: "https://github.com/ConductionNL/verhuis-service/blob/master/README.md"
+    apiDocumentation: "vs.zuid-drecht.nl"
 #
 #    features:
 #       - Very important feature
